Move human formatting functions under beets.util.units
